everything compiles and passes tests locally
