+++ tools/gfx/cpu/cpu-helper-functions.h
@@ -6,6 +6,6 @@ namespace gfx
 {
 using namespace Slang;
 
-Result SLANG_MCALL createCPUDevice(const IDevice::Desc* desc, IDevice** outDevice);
+
 
 } // namespace gfx
