+++ tools/gfx/debug-layer/debug-helper-functions.h
@@ -38,7 +38,7 @@ namespace debug
 #    define SLANG_FUNC_SIG "UnknownFunction"
 #endif
 
-thread_local const char* _currentFunctionName = nullptr;
+extern thread_local const char* _currentFunctionName;
 struct SetCurrentFuncRAII
 {
     SetCurrentFuncRAII(const char* funcName) { _currentFunctionName = funcName; }
@@ -56,10 +56,29 @@ char* _gfxDiagnoseFormat(
     size_t shortBufferSize, // Size of the initial buffer.
     List<char>& bufferArray, // A list for allocating a large buffer if needed.
     const char* format, // The format string.
-    TArgs... args);
+    TArgs... args)
+{
+    int length = sprintf_s(buffer, shortBufferSize, format, args...);
+    if (length < 0)
+        return buffer;
+    if (length > 255)
+    {
+        bufferArray.setCount(length + 1);
+        buffer = bufferArray.getBuffer();
+        sprintf_s(buffer, bufferArray.getCount(), format, args...);
+    }
+    return buffer;
+}
 
 template <typename... TArgs>
-void _gfxDiagnoseImpl(DebugMessageType type, const char* format, TArgs... args);
+void _gfxDiagnoseImpl(DebugMessageType type, const char* format, TArgs... args)
+{
+    char shortBuffer[256];
+    List<char> bufferArray;
+    auto buffer =
+        _gfxDiagnoseFormat(shortBuffer, sizeof(shortBuffer), bufferArray, format, args...);
+    getDebugCallback()->handleMessage(type, DebugMessageSource::Layer, buffer);
+}
 
 #define GFX_DIAGNOSE_ERROR(message)                                                                \
     _gfxDiagnoseImpl(                                                                              \
@@ -110,35 +129,11 @@ void _gfxDiagnoseImpl(DebugMessageType type, const char* format, TArgs... args);
                     : nullptr;                                                            \
     }
 
-SLANG_GFX_DEBUG_GET_INTERFACE_IMPL(Device)
-SLANG_GFX_DEBUG_GET_INTERFACE_IMPL_PARENT(BufferResource, Resource)
-SLANG_GFX_DEBUG_GET_INTERFACE_IMPL_PARENT(TextureResource, Resource)
-SLANG_GFX_DEBUG_GET_INTERFACE_IMPL(CommandBuffer)
-SLANG_GFX_DEBUG_GET_INTERFACE_IMPL(CommandQueue)
-SLANG_GFX_DEBUG_GET_INTERFACE_IMPL(Framebuffer)
-SLANG_GFX_DEBUG_GET_INTERFACE_IMPL(FramebufferLayout)
-SLANG_GFX_DEBUG_GET_INTERFACE_IMPL(InputLayout)
-SLANG_GFX_DEBUG_GET_INTERFACE_IMPL(RenderPassLayout)
-SLANG_GFX_DEBUG_GET_INTERFACE_IMPL(PipelineState)
-SLANG_GFX_DEBUG_GET_INTERFACE_IMPL(ResourceView)
-SLANG_GFX_DEBUG_GET_INTERFACE_IMPL(SamplerState)
-SLANG_GFX_DEBUG_GET_INTERFACE_IMPL(ShaderObject)
-SLANG_GFX_DEBUG_GET_INTERFACE_IMPL(ShaderProgram)
-SLANG_GFX_DEBUG_GET_INTERFACE_IMPL(Swapchain)
-SLANG_GFX_DEBUG_GET_INTERFACE_IMPL(TransientResourceHeap)
-SLANG_GFX_DEBUG_GET_INTERFACE_IMPL(QueryPool)
-SLANG_GFX_DEBUG_GET_INTERFACE_IMPL_PARENT(AccelerationStructure, ResourceView)
-SLANG_GFX_DEBUG_GET_INTERFACE_IMPL(Fence)
-SLANG_GFX_DEBUG_GET_INTERFACE_IMPL(ShaderTable)
-
-#undef SLANG_GFX_DEBUG_GET_INTERFACE_IMPL
-#undef SLANG_GFX_DEBUG_GET_INTERFACE_IMPL_PARENT
-
 // Utility conversion functions to get Debug* object or the inner object from a user provided
 // pointer.
 #define SLANG_GFX_DEBUG_GET_OBJ_IMPL(type)                                                   \
-    static Debug##type* getDebugObj(I##type* ptr) { return static_cast<Debug##type*>(ptr); } \
-    static I##type* getInnerObj(I##type* ptr)                                                \
+    inline Debug##type* getDebugObj(I##type* ptr) { return static_cast<Debug##type*>(ptr); } \
+    inline I##type* getInnerObj(I##type* ptr)                                                \
     {                                                                                        \
         if (!ptr) return nullptr;                                                            \
         auto debugObj = getDebugObj(ptr);                                                    \
@@ -169,8 +164,6 @@ SLANG_GFX_DEBUG_GET_OBJ_IMPL(QueryPool)
 SLANG_GFX_DEBUG_GET_OBJ_IMPL(AccelerationStructure)
 SLANG_GFX_DEBUG_GET_OBJ_IMPL(Fence)
 SLANG_GFX_DEBUG_GET_OBJ_IMPL(ShaderTable)
-
-#undef SLANG_GFX_DEBUG_GET_OBJ_IMPL
 
 void validateAccelerationStructureBuildInputs(
     const IAccelerationStructure::BuildInputs& buildInputs);
